+++ docs/examples/parse_directory.py
@@ -1,0 +1,3 @@
+from ms3 import Parse
+p = Parse('..')
+print(p)