+++ docs/examples/parse_directory_mscx.py
@@ -1,0 +1,3 @@
+from ms3 import Parse
+p = Parse('..', file_re="mscx$", key='ms3')
+print(p)